% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkmissing-package.R
\docType{package}
\name{odkmissing-package}
\alias{odkmissing}
\alias{odkmissing-package}
\title{odkmissing: herramientas para flags de missing con ODK}
\description{
Herramientas para importar un XLSForm ODK y generar flags de missings
}
\author{
\strong{Maintainer}: Julián López \email{julianlopces@gmail.com}

}
\keyword{internal}
