% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okrmissing-package.R
\docType{package}
\name{okrmissing-package}
\alias{okrmissing}
\alias{okrmissing-package}
\title{okrmissing: herramientas para flags de missing con ODK}
\description{
Herramientas para importar un XLSForm ODK y generar flags de missings
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
