% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okrmissing-package.R
\docType{package}
\name{okrmissing}
\alias{okrmissing-package}
\alias{okrmissing}
\title{okrmissing: Quality Control de missings de formularios a partir de ODK}
\description{
Herramientas para importar un XLSForm ODK y generar flags de missings
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
